eliminate need for a common rust lib
