+++ sources/sub-projects/ayfstarter/sources/Components/AudioNodes/ayfAuNStarter/src/CayfAuNStarter.h
@@ -27,7 +27,6 @@
 extern "C"
 {
 #ifdef JVX_USE_RUST_CORE_LIB
-#include "ayfstarterlibrs_common.h"
 #include "ayfstarterlibrs.h"
 #ifdef JVX_EXTERNAL_CALL_ENABLED
 #include "ayfstarterlibdbgrs.h"
