--- sources/sub-projects/ayfstarter/sources/Libraries/rust/ayfstarterlib-common/.gitignore
@@ -1,1 +1,0 @@
-/target
