--- sources/sub-projects/ayfstarter/sources/Libraries/rust/ayfstarterlib-common/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "ayfstarterlibrs_common"
-publish = false
-edition = "2021"
-
-[features]
-default = []
-
-[dependencies]
-jvx_dsp = { path = "../../../../../../jvxLibraries/jvx-dsp-rust" }
