--- sources/sub-projects/ayfstarter/sources/Libraries/rust/ayfstarterlib-common/cbindgen.toml
@@ -1,17 +1,0 @@
-# The language to output bindings in
-#
-# possible values: "C", "C++", "Cython"
-#
-# default: "C++"
-language = "C"
-
-# Whether to add a `#pragma once` guard
-# default: doesn't emit a `#pragma once`
-pragma_once = true
-
-# A list of headers to #include (with quotes)
-# default: []
-includes = ["jvx_rust_audyoflo_api.h"]
-
-# Add the keyword "struct"
-style = "tag"
