+++ sources/sub-projects/ayfstarter/sources/Libraries/rust/ayfstarterlib/Cargo.toml
@@ -4,7 +4,7 @@ publish = false
 edition = "2021"
 
 [lib]
-crate-type = ["staticlib"]
+crate-type = ["lib", "staticlib"]
 
 [features]                                                                                                  
 default = []
@@ -12,5 +12,3 @@ jvx-system-use-data-format-float = ["jvx_dsp/jvx-system-use-data-format-float"]
 
 [dependencies]
 jvx_dsp = { path = "../../../../../../jvxLibraries/jvx-dsp-rust" }
-ayfstarterlibrs_common = { path = "../ayfstarterlib-common" }
-
