+++ sources/sub-projects/ayfstarter/sources/Libraries/rust/ayfstarterlib_debug/Cargo.toml
@@ -13,4 +13,4 @@ jvx-allow-profiler-data-macros = []
 
 [dependencies]
 jvx_dsp = { path = "../../../../../../jvxLibraries/jvx-dsp-rust" }
-ayfstarterlibrs_common = { path = "../ayfstarterlib-common" }
+ayfstarterlibrs = { path = "../ayfstarterlib" }
